replace all occurences of piece_id as short with int32_t

+++ src/Player.hpp
@@ -21,23 +21,23 @@ class Player final
     Player(const Player &) = delete;
     Player &operator=(const Player &) = delete;
     void receivePiece(PiecePtr &piecePtr);
-    void losePiece(const short targetId);
-    [[nodiscard]] const std::unordered_map<short, chk::PiecePtr> &getOwnPieces() const;
-    void showMyHunters(const std::unordered_set<short> &hunterPieces) const;
+    void losePiece(const int32_t targetId);
+    [[nodiscard]] const std::unordered_map<int32_t, chk::PiecePtr> &getOwnPieces() const;
+    void showMyHunters(const std::unordered_set<int32_t> &hunterPieces) const;
     void clearBasket();
     [[nodiscard]] size_t getPieceCount() const;
     [[nodiscard]] const std::string &getName() const;
     [[nodiscard]] PlayerType getPlayerType() const;
-    [[nodiscard]] bool hasThisPiece(const short pieceId) const;
-    [[nodiscard]] bool movePiece(const short pieceId, const sf::Vector2f &destPos);
-    [[nodiscard]] bool captureEnemyWith(const short pieceId, const sf::Vector2f &destPos);
+    [[nodiscard]] bool hasThisPiece(const int32_t pieceId) const;
+    [[nodiscard]] bool movePiece(const int32_t pieceId, const sf::Vector2f &destPos);
+    [[nodiscard]] bool captureEnemyWith(const int32_t pieceId, const sf::Vector2f &destPos);
     bool operator==(const Player &other) const;
 
   private:
     // name of this player (RED or BLACK)
     std::string name;
     // hashmap of my PieceId -> its Pointer
-    std::unordered_map<short, chk::PiecePtr> basket;
+    std::unordered_map<int32_t, chk::PiecePtr> basket;
 };
 
 } // namespace chk
